fix(frontend:eslint): fix eslint errors
