+++ frontend/pages/dashboard.vue
@@ -2,4 +2,5 @@
 </script>
 
 <template>
+    <div />
 </template>
