+++ frontend/pages/dashboardHost.vue
@@ -1,5 +1,6 @@
-<script setup>
+<script setup lang="ts">
 </script>
 
 <template>
+    <div />
 </template>
